# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Chronos API",
    "version": "0.9.0",
    "description": "Self-hostable workflow orchestration engine — REST API for managing workflows, executions, API keys, and webhooks.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/gautamneeraj88/chronos/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://your-domain.com/api",
      "description": "Production (behind reverse proxy)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT from POST /auth/login or an API key from the dashboard"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "member"
            ]
          },
          "orgId": {
            "type": "string"
          }
        }
      },
      "WorkflowStep": {
        "type": "object",
        "required": [
          "name",
          "type",
          "activity"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "charge-card"
          },
          "type": {
            "type": "string",
            "enum": [
              "activity"
            ],
            "description": "Must be \"activity\""
          },
          "activity": {
            "type": "string",
            "example": "chargeCard"
          },
          "compensation": {
            "type": "string",
            "nullable": true
          },
          "retries": {
            "type": "integer",
            "default": 0
          },
          "timeoutMs": {
            "type": "integer",
            "default": 30000
          }
        }
      },
      "CreateWorkflowInput": {
        "type": "object",
        "required": [
          "name",
          "steps"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "order-processing"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowStep"
            }
          }
        }
      },
      "Workflow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          },
          "orgId": {
            "type": "string"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowStep"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Execution": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "workflowId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "RUNNING",
              "COMPLETED",
              "COMPENSATING",
              "COMPENSATED",
              "COMPENSATION_FAILED"
            ]
          },
          "currentStepIndex": {
            "type": "integer"
          },
          "input": {
            "type": "object",
            "additionalProperties": true
          },
          "output": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "triggeredBy": {
            "type": "string"
          },
          "orgId": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "ExecutionEvent": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "EXECUTION_STARTED",
              "STEP_STARTED",
              "STEP_IN_FLIGHT",
              "STEP_COMPLETED",
              "STEP_FAILED",
              "EXECUTION_COMPLETED",
              "COMPENSATION_STARTED",
              "STEP_COMPENSATED",
              "EXECUTION_COMPENSATED",
              "COMPENSATION_FAILED"
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Auth",
      "description": "Authentication and user management"
    },
    {
      "name": "Workflows",
      "description": "Workflow definition management"
    },
    {
      "name": "Executions",
      "description": "Execution triggering and status"
    },
    {
      "name": "API Keys",
      "description": "API key management"
    },
    {
      "name": "Webhooks",
      "description": "Webhook registration"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Exchange email + password for a JWT",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "admin@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "your-password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing email or password"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Return the current authenticated user",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Refresh JWT before expiry",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "New token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Create a new user (admin only)",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "member"
                    ],
                    "default": "member"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "description": "Missing fields"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Admin role required"
          }
        }
      }
    },
    "/auth/users": {
      "get": {
        "summary": "List all users in the org (admin only)",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of users"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Admin role required"
          }
        }
      }
    },
    "/auth/users/{id}": {
      "delete": {
        "summary": "Delete a user by ID (admin only)",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Admin role required"
          }
        }
      }
    },
    "/api-keys": {
      "get": {
        "summary": "List all API keys for the org",
        "tags": [
          "API Keys"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of API keys (secret not included)"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create a new API key (admin only)",
        "tags": [
          "API Keys"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "ci-pipeline"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created — key field only present in this response"
          },
          "400": {
            "description": "Name is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Admin role required"
          }
        }
      }
    },
    "/api-keys/{id}": {
      "delete": {
        "summary": "Revoke an API key (admin only)",
        "tags": [
          "API Keys"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "API key revoked"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Admin role required"
          }
        }
      }
    },
    "/executions": {
      "get": {
        "summary": "List executions for the org",
        "tags": [
          "Executions"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "RUNNING",
                "COMPLETED",
                "COMPENSATING",
                "COMPENSATED",
                "COMPENSATION_FAILED"
              ]
            },
            "description": "Filter by execution status"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of executions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Execution"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Trigger an execution by workflowId",
        "tags": [
          "Executions"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "workflowId"
                ],
                "properties": {
                  "workflowId": {
                    "type": "string"
                  },
                  "input": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Execution started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Execution"
                }
              }
            }
          },
          "400": {
            "description": "workflowId is required"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/executions/{id}": {
      "get": {
        "summary": "Get execution status",
        "tags": [
          "Executions"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Execution object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Execution"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/executions/{id}/events": {
      "get": {
        "summary": "Get the full append-only event log for an execution",
        "tags": [
          "Executions"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ordered array of execution events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExecutionEvent"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/webhooks": {
      "get": {
        "summary": "List all webhooks for the org",
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of webhooks"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Register a new webhook",
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "url",
                  "events"
                ],
                "properties": {
                  "url": {
                    "type": "string",
                    "example": "https://example.com/hooks/chronos"
                  },
                  "events": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "execution.started",
                        "execution.completed",
                        "execution.compensated",
                        "execution.failed",
                        "step.completed",
                        "step.failed"
                      ]
                    }
                  },
                  "secret": {
                    "type": "string",
                    "description": "Optional HMAC-SHA256 signing secret"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook registered"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/webhooks/{id}": {
      "delete": {
        "summary": "Remove a webhook",
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Webhook deleted"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/workflows": {
      "post": {
        "summary": "Register a new workflow definition",
        "tags": [
          "Workflows"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workflow created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workflow"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "List all workflow definitions for the org",
        "tags": [
          "Workflows"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of workflow definitions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Workflow"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/workflows/{id}": {
      "get": {
        "summary": "Get a workflow definition by ID",
        "tags": [
          "Workflows"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow definition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workflow"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/workflows/{id}/executions": {
      "post": {
        "summary": "Trigger an execution for a workflow",
        "tags": [
          "Workflows"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Workflow ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "input": {
                    "type": "object",
                    "additionalProperties": true,
                    "example": {
                      "orderId": "ORD-001",
                      "amount": 99.99
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Execution started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Execution"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Workflow not found"
          }
        }
      }
    }
  }
}